+++ src/IntegrationTests/Models/Movie.cs
