+++ src/IntegrationTests/Models/TimeStamp.cs
