+++ src/IntegrationTests/Queries/delete_movies.cypher
