+++ src/Neo4jMapper/RecordExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Neo4j.Driver.V1;
 
-namespace Neo4j.Driver.Extensions
+namespace Neo4jMapper
 {
     public static class RecordExtensions
     {
