+++ src/Neo4jMapper/SessionExtensions.cs
@@ -5,7 +5,7 @@ using System.Reflection;
 using Neo4j.Driver.V1;
 using ServiceStack;
 
-namespace Neo4j.Driver.Extensions
+namespace Neo4jMapper
 {
     public static class SessionExtensions
     {
