+++ src/Neo4jMapper/StatementResultExtensions.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using Neo4j.Driver.V1;
 
-namespace Neo4j.Driver.Extensions
+namespace Neo4jMapper
 {
     public static class StatementResultExtensions
     {
